:construction: feat (content) : wip for return error/doc

+++ src/controller/content.controller.ts
@@ -9,7 +9,6 @@ export class ContentController {
 
     @Get(':id')
     async getContent(@Param('id') id): Promise<object> {
-
         if (!isMongoId(id)) {
             throw new NotFoundException('This id dosen\'t exist.');
         }
@@ -18,24 +17,40 @@ export class ContentController {
 
         // if no data
         if (data.length === 0) {
-            throw new NotFoundException('This id dosen\'t exist.');
+            throw new NotFoundException('No data.');
         }
 
         return {status: 'ok', data};
     }
 
     @Post('add')
-    setContent(@Body() body: DtoContent): object {
-        return this.contentService.setContent(body);
+    async setContent(@Body() body: DtoContent): Promise<object> {
+        const res = await this.contentService.setContent(body);
+        if (res['errors']) {
+            return {status : 'error', message : res['errors']}
+        } else {
+            return {status : 'ok', message : res}
+        }
     }
 
     @Put(':id')
     updateContent(@Param('id') id, @Body() body: DtoContent): object {
+        if (!isMongoId(id)) {
+            throw new NotFoundException('This id dosen\'t exist.');
+        }
         return this.contentService.updateContent(id, body);
     }
 
     @Delete(':id')
-    deleteContent(@Param('id') id): string {
-        return this.contentService.deleteContent(id);
+    async deleteContent(@Param('id') id): Promise<object> {
+        if (!isMongoId(id)) {
+            throw new NotFoundException('This id dosen\'t exist.');
+        }
+        const res = await this.contentService.deleteContent(id);
+        if ( res['deletedCount'] <= 0) {
+            return {status : 'error', message : res}
+        } else {
+            return {status : 'ok', message : res}
+        }
     }
 }
